refactor: textview handler를 커스텀 텍스트뷰로 대체
